+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # reef-knot
 
+## 0.5.1
+
+### Patch Changes
+
+- Improve the modal's scrollbar
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes